+++ python/paddle/jit/dy2static/ast_utils.py
@@ -34,6 +34,7 @@ def ast_to_source_code(ast_node):
         ast_node = gast.gast_to_ast(ast_node)
 
     if sys.version_info >= (3, 9):
+        ast.fix_missing_locations(ast_node)
         return ast.unparse(ast_node)
 
     # Do not wrap lines even if they are too long
