+++ python/paddle/jit/dy2static/utils.py
@@ -38,7 +38,7 @@ from paddle.base.layer_helper import LayerHelper
 from paddle.base.wrapped_decorator import signature_safe_contextmanager
 from paddle.utils import gast
 
-from .transformers import ast_to_source_code
+from .ast_utils import ast_to_source_code
 from .utils_helper import (  # noqa: F401
     DYGRAPH_MODULE_PREFIX,
     DYGRAPH_TO_STATIC_MODULE_PREFIX,
